Add RocksDBTimestampedSegmentedBytesStoreTest Fix bug revealed by new test
